+++ CompanyService/src/main/java/by/ares/company_service/dto/request/UpdateCarRequest.java
@@ -10,13 +10,13 @@ public record UpdateCarRequest(
         String carName,
 
         @Positive(message = "Fuel consumption cannot be negative")
-        Integer fuelConsumption,
+        Float fuelConsumption,
 
         @Positive(message = "Action fuel consumption cannot be negative")
-        Integer actionFuelConsumption,
+        Float actionFuelConsumption,
 
         @Positive(message = "Oil consumption cannot be negative")
-        Integer oilConsumption,
+        Float oilConsumption,
 
         @NotNull(message = "Car type must be provided")
         CarType carType
